Splitting get container and vault tests into separate files
